# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Restful Provisioning Protocol (RPP) ",
    "description": "A Modern Domain Registry API",
    "version": "0.0.1"
  },
  "paths": {
    "/entities/": {
      "post": {
        "tags": [
          "Entities"
        ],
        "summary": "Create Entity",
        "description": "Create a new entity",
        "operationId": "do_create_entities__post",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{entity_id}": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "Get Entity Info",
        "operationId": "do_info_entities__entity_id__get",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "rpp-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Authorization"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities"
        ],
        "summary": "Delete Entity",
        "operationId": "do_delete_entities__entity_id__delete",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response",
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Entities"
        ],
        "summary": "Update Entity",
        "operationId": "do_update_entities__entity_id__patch",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/entities/{entity_id}/processes/transfers": {
      "post": {
        "tags": [
          "Entities"
        ],
        "summary": "Start Entity Transfer",
        "operationId": "do_start_transfer_entities__entity_id__processes_transfers_post",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "rpp-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Authorization"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "Query Transfer Status",
        "operationId": "do_query_transfer_entities__entity_id__processes_transfers_get",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "rpp-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Authorization"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{entity_id}/processes/transfers/rejection": {
      "put": {
        "tags": [
          "Entities"
        ],
        "summary": "Reject Entity Transfer",
        "operationId": "do_reject_transfer_entities__entity_id__processes_transfers_rejection_put",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "rpp-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Authorization"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{entity_id}/processes/transfers/cancellation": {
      "put": {
        "tags": [
          "Entities"
        ],
        "summary": "Cancel Entity Transfer",
        "operationId": "do_cancel_transfer_entities__entity_id__processes_transfers_cancellation_put",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "rpp-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Authorization"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthInfoModel": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "roid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roid"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "AuthInfoModel"
      },
      "BaseResponseModel": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type"
          },
          "result": {
            "items": {
              "$ref": "#/components/schemas/ResultModel"
            },
            "type": "array",
            "title": "Result"
          },
          "messages": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageQueueModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "resData": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Resdata"
          },
          "extension": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Extension"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "BaseResponseModel"
      },
...
      "EntityCreateRequest": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type"
          },
          "card": {
            "$ref": "#/components/schemas/Card"
          },
          "authInfo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthInfoModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "card"
        ],
        "title": "EntityCreateRequest"
      },
...
      "EntityUpdateRequest": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type"
          },
          "add": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityUpdateAddOrRemove"
              },
              {
                "type": "null"
              }
            ]
          },
          "remove": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityUpdateAddOrRemove"
              },
              {
                "type": "null"
              }
            ]
          },
          "change": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityUpdateChange"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "EntityUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}